rnpkeys: Compile rnpkeys to library for testing
