+++ src/lib/Makefile.am
@@ -1,13 +1,11 @@
 ## $NetBSD: Makefile.am,v 1.10 2010/11/29 06:21:40 agc Exp $
 
-AM_CFLAGS		= $(WARNCFLAGS) $(BOTAN_INCLUDES) $(JSONC_INCLUDES)
-
-AM_LDFLAGS              = $(BOTAN_LDFLAGS) $(JSONC_LDFLAGS)
-
+AM_CFLAGS			= $(WARNCFLAGS) $(BOTAN_INCLUDES) $(JSON_CFLAGS)
+AM_LDFLAGS          = $(BOTAN_LDFLAGS)
 lib_LTLIBRARIES		= librnp.la
-
 librnp_la_CPPFLAGS	= -I$(top_srcdir)/include
-
+librnp_la_LIBADD	= $(BOTAN_LIBS) $(JSON_LIBS)
+dist_man_MANS		= librnp.3
 librnp_la_SOURCES	= \
 	bn.c \
 	bufgap.c \
@@ -37,7 +35,3 @@ librnp_la_SOURCES	= \
 	symmetric.c \
 	validate.c \
 	writer.c
-
-librnp_la_LIBADD	= $(BOTAN_LIBS) $(JSONC_LIBS)
-
-dist_man_MANS		= librnp.3
