+++ src/rnpkeys/Makefile.am
@@ -1,12 +1,13 @@
-AM_CFLAGS		= $(WARNCFLAGS)
+AM_CFLAGS   		    = $(WARNCFLAGS)
+dist_man_MANS           = rnpkeys.1
 
-bin_PROGRAMS		= rnpkeys
+lib_LTLIBRARIES         = librnpkeys.la
+librnpkeys_la_CPPFLAGS  = -I$(top_srcdir)/include $(JSONC_INCLUDES)
+librnpkeys_la_LIBADD    = ../lib/librnp.la
+librnpkeys_la_SOURCES   = \
+    rnpkeys.c \
+    tui.c
 
-rnpkeys_SOURCES	=   rnpkeys.c \
-                    tui.c
-
-rnpkeys_CPPFLAGS	= -I$(top_srcdir)/include $(JSONC_INCLUDES)
-
-rnpkeys_LDADD	= ../lib/librnp.la
-
-dist_man_MANS		= rnpkeys.1
+bin_PROGRAMS            = rnpkeys
+rnpkeys_SOURCES         = main.c
+rnpkeys_LDADD           = librnpkeys.la
