+++ src/rnpv/Makefile.am
@@ -1,17 +1,12 @@
-AM_CFLAGS		= $(WARNCFLAGS) $(JSONC_INCLUDES)
-
-bin_PROGRAMS		= rnpv
-
-rnpv_SOURCES	= \
-	b64.c \
-	bufgap.c \
-	libverify.c \
-	main.c \
-	misc.c \
-	pgpsum.c
-
-rnpv_CPPFLAGS	= -I$(top_srcdir)/include -I$(top_srcdir)/src/lib
-
-rnpv_LDADD	= ../lib/librnp.la $(JSONC_LIBS)
-
-dist_man_MANS		= rnpv.1
+AM_CFLAGS		= $(WARNCFLAGS) $(JSON_CFLAGS)
+bin_PROGRAMS	= rnpv
+rnpv_CPPFLAGS   = -I$(top_srcdir)/include -I$(top_srcdir)/src/lib
+rnpv_LDADD	    = ../lib/librnp.la $(JSON_LIBS)
+dist_man_MANS	= rnpv.1
+rnpv_SOURCES    = \
+    b64.c \
+    bufgap.c \
+    libverify.c \
+    main.c \
+    misc.c \
+    pgpsum.c
